# src/docs/api.docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lowlesses User API",
    "description": "Тут описаны методы для взаимодействия с сущностью пользователя",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
    "/create-user": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Создать пользователя",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/update-user": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Создать пользователя",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstname",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastname",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "patronomic",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "telegram",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "vk",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "whatsapp",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "discord",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/login": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Авторизоваться",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/check-token": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Проверить токен на валидность и время действия",
        "parameters": [
          {
            "name": "token",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/set-user-hash": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Установить хэш пользователя, который пригласил в игру",
        "parameters": [
          {
            "name": "invitedHash",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/set-user-role": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/user"],
        "description": "Установить роль",
        "parameters": [
          {
            "name": "role",
            "in": "formData",
            "required": true,
            "type": "string",
            "enum": ["USER", "ADMIN", "ROOT", "MODERATOR"]
          }
        ],
        "responses": {}
      }
    },
...
    "/create": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/news"],
        "description": "Создать новость",
        "parameters": [
          {
            "name": "titleRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "titleEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "bodyRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "bodyEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "descriptionRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "descriptionEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "isShow",
            "in": "formData",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "avatar",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "tags",
            "in": "formData",
            "required": true,
            "type": "array",
            "items": { "type": "string" }
          }
        ],
        "responses": {}
      }
    },
    "/update-news/:slug": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/news"],
        "description": "Изменить новость",
        "parameters": [
          {
            "name": "titleRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "titleEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "bodyRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "bodyEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "descriptionRU",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "descriptionEN",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "isShow",
            "in": "formData",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "avatar",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "tags",
            "in": "formData",
            "required": true,
            "type": "array",
            "items": { "type": "string" }
          },
          {
            "name": "slug",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "formData",
            "required": false,
            "type": "number"
          },
          {
            "name": "autorId",
            "in": "formData",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {}
      }
    },
...
    "/get-news/:slug": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/news"],
        "description": "Получить новость по ссылке",
        "parameters": [
          {
            "name": "locale",
            "in": "formData",
            "required": true,
            "type": "string",
            "enum": ["ru-RU", "en-EN"]
          }
        ],
        "responses": {}
      }
    },
    "/get-many": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "index",
        "tags": ["/news"],
        "description": "Получить список новостей",
        "parameters": [
          {
            "name": "limit",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "locale",
            "in": "formData",
            "required": true,
            "type": "string",
            "enum": ["ru-RU", "en-EN"]
          }
        ],
        "responses": {}
      }
    },
...
  }
}